FashionClassifierUI/requirements.txt
absl-py==1.3.0
altair==4.2.0
astunparse==1.6.3
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
decorator==5.1.1
entrypoints==0.4
flatbuffers==22.12.6
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
importlib-metadata==5.1.0
Jinja2==3.1.2
jsonschema==4.17.3
keras==2.11.0
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
protobuf==3.20.0
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.16.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
watchdog==2.2.0
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
